+++ ReleaseNotes.md
@@ -10,3 +10,4 @@
 
 ### New in 0.0.4
 * Kernel is now resolved as a singleton from container.
+* Moved the Lake console build log to Lunt.